+++ src/main/java/crm/common/model/Product.java
@@ -8,8 +8,8 @@ import java.util.List;
 @Entity(tableName = "Product")
 public class Product {
     @Key
-    @Column(name = "ProductID", type = "BIGINT")
-    private Long productID;
+    @Column(name = "ProductID", type = "INT")
+    private Integer productID;
 
     @Column(name = "ProductName", length = 150, nullable = false)
     private String productName;
@@ -20,8 +20,8 @@ public class Product {
     @Column(name = "ProductDescription", length = 255)
     private String productDescription;
 
-    @Column(name = "TypeID", type = "BIGINT", nullable = false)
-    private Long typeID;
+    @Column(name = "TypeID", type = "INT", nullable = false)
+    private Integer typeID;
 
     @ManyToOne(joinColumn = "TypeID")
     private LazyReference<Type> type;
@@ -29,11 +29,11 @@ public class Product {
     @OneToMany(mappedBy = "productID", joinColumn = "ProductID")
     private List<ProductSpecification> productSpecifications;
 
-    public Long getProductID() {
+    public Integer getProductID() {
         return productID;
     }
 
-    public void setProductID(Long productID) {
+    public void setProductID(Integer productID) {
         this.productID = productID;
     }
 
@@ -61,11 +61,11 @@ public class Product {
         this.productDescription = productDescription;
     }
 
-    public Long getTypeID() {
+    public Integer getTypeID() {
         return typeID;
     }
 
-    public void setTypeID(Long typeID) {
+    public void setTypeID(Integer typeID) {
         this.typeID = typeID;
     }
